+++ src/FileArray.php
@@ -44,7 +44,8 @@ class FileArray
      */
     public function __construct($filename, array $data = [])
     {
-        is_writable_or_fail(is_file($filename) ? $filename : dirname($filename));
+        $target = is_file($filename) ? $filename : dirname($filename);
+        is_true_or_fail(is_writable($target), sprintf('File or directory `%s` is not writable', rtr($target)));
 
         $this->filename = $filename;
         $this->data = $data ?: $this->read();
@@ -189,6 +190,6 @@ class FileArray
      */
     public function write()
     {
-        return (bool)file_put_contents($this->filename, serialize($this->data));
+        return create_file($this->filename, serialize($this->data));
     }
 }